refactor(models): extract CuratorProposedTerm abstract base for FocusArea and Skill
Both models had identical structure (name, reviewed, ordering, __str__).
Shared fields now live in an abstract base class so future taxonomy
terms can reuse the pattern.
