+++ backend/terramedic/organizations/models/focus_area.py
@@ -1,12 +1,6 @@
-from django.db import models
+from terramedic.organizations.models.base import CuratorProposedTerm
 
 
-class FocusArea(models.Model):
-    name = models.CharField(max_length=100, unique=True)
-    reviewed = models.BooleanField(default=False)
-
-    class Meta:
-        ordering = ["name"]
-
-    def __str__(self) -> str:
-        return self.name
+class FocusArea(CuratorProposedTerm):
+    class Meta(CuratorProposedTerm.Meta):
+        pass
