+++ backend/terramedic/organizations/models/skill.py
@@ -1,12 +1,6 @@
-from django.db import models
+from terramedic.organizations.models.base import CuratorProposedTerm
 
 
-class Skill(models.Model):
-    name = models.CharField(max_length=100, unique=True)
-    reviewed = models.BooleanField(default=False)
-
-    class Meta:
-        ordering = ["name"]
-
-    def __str__(self) -> str:
-        return self.name
+class Skill(CuratorProposedTerm):
+    class Meta(CuratorProposedTerm.Meta):
+        pass
